Remove unnecessary dtype='object' in multivariate filters

+++ ITMO_FS/filters/multivariate/DISRwithMassive.py
@@ -131,7 +131,7 @@ class DISRWithMassive(DataChecker):
         """
 
         if type(X) is np.ndarray:
-            return X[:, self.selected_features.astype(int)]
+            return X[:, self.selected_features]
         else:
             return X[self.selected_features]
